refactor(cli): decouple mine.go + handoff.go from rpi symbols (ag-362v #decouple)
Prerequisite for removing the ao rpi / ao evolve CLI engines (ag-llni,
soc-1gbpz). Verified that the ONLY non-rpi/evolve production callers of rpi
symbols are mine.go (RPIC2Event + loadRPIC2Events) and handoff.go
(shellQuote). Relocate both into files that survive the rpi*.go deletion
glob, with ZERO behavior change:

- shell_quote.go: POSIX single-quote helper, moved out of rpi_parallel.go
  (inlined body; drops the internal/rpi.ShellQuote indirection). handoff.go
  and rpi_parallel.go both keep using it.
- c2_event_reader.go: RPIC2Event type + loadRPIC2Events reader, moved out of
  rpi_c2_events.go (events.jsonl path inlined, mirroring
  internal/rpi.RPIRunRegistryDir). mine.go unchanged; the rpi-internal writer
  appendRPIC2Event stays in rpi_c2_events.go and is deleted with the engine.

After this, no code outside rpi*.go/evolve*.go references any rpi-internal
symbol — the engine deletion (next PR) is unblocked.

Verified: go build ./... + go vet clean; 3 new reader tests + 21 existing
handoff/mine/c2/shellQuote tests pass; rg confirms mine.go/handoff.go carry
no rpi-internal symbol refs.

Closes-scenario: ag-362v#decouple
Bounded-context: BC5-Runtime
Evidence: cli/cmd/ao/c2_event_reader.go
